Defining interfaces for IAlmostEquatableComparator and IAlmostComparableComparator; Implementing a 2D line class; Updating the BPlusTree for using big integer for avoiding an overflow for large trees.

+++ DevEnvironment/DataStructures/src/main/java/datastructures/bplustree/interfaces/IBPlusTreeMetrics.java
@@ -2,6 +2,7 @@ package datastructures.bplustree.interfaces;
 
 import base.interfaces.IUnaryComparator;
 import datastructures.tree.interfaces.ITreeMetrics;
+import java.math.BigInteger;
 
 /**
  * The ITreeMetrics interface defines metrics of a tree.
@@ -10,5 +11,5 @@ public interface IBPlusTreeMetrics extends ITreeMetrics, IUnaryComparator<IBPlus
     /**
      * Gets the capacity in bytes of a tree.
      */
-    long getCapacityInBytes();
+    BigInteger getCapacityInBytes();
 }
